+++ app/models/utility/scenario/units/infantry_weapons.rb
@@ -16,8 +16,8 @@ module Utility
             key = %i[c n y o]
             [
               ["alm", "M1915 Chauchat", 15, { sn: 1 }], ["alm", "Bren LMG", 35],
-              ["alm", "Vickers MG", 12], ["alm", "Colt M/29", 17], ["alm", "MG 08/15", 24],
-              ["alm", "rkm wz. 28", 28], ["alm", "ZB vz. 30", 26],
+              ["alm", "Vickers MG", 12], ["alm", "Colt M/29", 17], ["alm", "12.7mm HMG", 17],
+              ["alm", "MG 08/15", 24], ["alm", "rkm wz. 28", 28], ["alm", "ZB vz. 30", 26],
 
               ["axm", "MG 08/15", 24], ["axm", "MG 30", 31], ["axm", "MG 34", 36],
               ["axm", "MG 42", 42], ["axm", "ZB vz. 26", 26], ["axm", "ZB vz. 30", 26],